CoursesPortalMentoringApp/src/app/core/header/header.component.ts
import { Component, OnInit } from '@angular/core';
import { Router } from '@angular/router';
import { AuthService } from '../auth/auth.service';


@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.css']
})
export class HeaderComponent implements OnInit {

  loginMessage = 'User Login';
  logoutMessage = 'log out';

  constructor(public authService: AuthService, private router: Router) {
  }

  ngOnInit() {
  }

  logout() {
    this.authService.logout();
    this.goToLoginPage();
  }

  goToLoginPage() {
    this.router.navigate(['/login']);
  }

}
